+++ src/core/loop/loopCreate.mjs
@@ -1,6 +1,6 @@
 import { createElement, elementChildren, showWarning } from '../../shared/utils.mjs';
 
-export default function loopCreate(slideRealIndex) {
+export default function loopCreate(slideRealIndex, initial) {
   const swiper = this;
   const { params, slidesEl } = swiper;
   if (!params.loop || (swiper.virtual && swiper.params.virtual.enabled)) return;
@@ -13,7 +13,22 @@ export default function loopCreate(slideRealIndex) {
     });
   };
 
+  const clearBlankSlides = () => {
+    const slides = elementChildren(slidesEl, `.${params.slideBlankClass}`);
+
+    slides.forEach((el) => {
+      el.remove();
+    });
+    if (slides.length > 0) {
+      swiper.recalcSlides();
+      swiper.updateSlides();
+    }
+  };
+
   const gridEnabled = swiper.grid && params.grid && params.grid.rows > 1;
+  if (params.loopAddBlankSlides && (params.slidesPerGroup > 1 || gridEnabled)) {
+    clearBlankSlides();
+  }
 
   const slidesPerGroup = params.slidesPerGroup * (gridEnabled ? params.grid.rows : 1);
 
@@ -58,5 +73,9 @@ export default function loopCreate(slideRealIndex) {
     initSlides();
   }
 
-  swiper.loopFix({ slideRealIndex, direction: params.centeredSlides ? undefined : 'next' });
+  swiper.loopFix({
+    slideRealIndex,
+    direction: params.centeredSlides ? undefined : 'next',
+    initial,
+  });
 }