+++ src/core/slide/slideToLoop.mjs
@@ -21,9 +21,9 @@ export default function slideToLoop(index = 0, speed, runCallbacks = true, inter
       let targetSlideIndex;
       if (gridEnabled) {
         const slideIndex = newIndex * swiper.params.grid.rows;
-        targetSlideIndex = swiper.slides.filter(
+        targetSlideIndex = swiper.slides.find(
           (slideEl) => slideEl.getAttribute('data-swiper-slide-index') * 1 === slideIndex,
-        )[0].column;
+        ).column;
       } else {
         targetSlideIndex = swiper.getSlideIndexByData(newIndex);
       }
@@ -70,9 +70,9 @@ export default function slideToLoop(index = 0, speed, runCallbacks = true, inter
 
       if (gridEnabled) {
         const slideIndex = newIndex * swiper.params.grid.rows;
-        newIndex = swiper.slides.filter(
+        newIndex = swiper.slides.find(
           (slideEl) => slideEl.getAttribute('data-swiper-slide-index') * 1 === slideIndex,
-        )[0].column;
+        ).column;
       } else {
         newIndex = swiper.getSlideIndexByData(newIndex);
       }