+++ src/modules/thumbs/thumbs.mjs
@@ -49,7 +49,12 @@ export default function Thumb({ swiper, extendParams, on }) {
     if (initialized) return false;
     initialized = true;
     const SwiperClass = swiper.constructor;
+
     if (thumbsParams.swiper instanceof SwiperClass) {
+      if (thumbsParams.swiper.destroyed) {
+        initialized = false;
+        return false;
+      }
       swiper.thumbs.swiper = thumbsParams.swiper;
       Object.assign(swiper.thumbs.swiper.originalParams, {
         watchSlidesProgress: true,
@@ -125,9 +130,9 @@ export default function Thumb({ swiper, extendParams, on }) {
       let newThumbsIndex;
       let direction;
       if (thumbsSwiper.params.loop) {
-        const newThumbsSlide = thumbsSwiper.slides.filter(
+        const newThumbsSlide = thumbsSwiper.slides.find(
           (slideEl) => slideEl.getAttribute('data-swiper-slide-index') === `${swiper.realIndex}`,
-        )[0];
+        );
         newThumbsIndex = thumbsSwiper.slides.indexOf(newThumbsSlide);
 
         direction = swiper.activeIndex > swiper.previousIndex ? 'next' : 'prev';