+++ src/types/swiper-options.d.ts
@@ -210,7 +210,7 @@ export interface SwiperOptions {
    *
    * @default 'slide'
    */
-  effect?: 'slide' | 'fade' | 'cube' | 'coverflow' | 'flip' | 'creative' | 'cards' | string;
+  effect?: 'slide' | 'fade' | 'cube' | 'coverflow' | 'flip' | 'creative' | 'cards' | (string & {});
 
   /**
    * Fire Transition/SlideChange/Start/End events on swiper initialization.
@@ -741,7 +741,7 @@ export interface SwiperOptions {
    *
    * @default 'window'
    */
-  breakpointsBase?: 'window' | 'container';
+  breakpointsBase?: 'window' | 'container' | CSSSelector;
 
   // Observer
   /**